Conflicts:
	README.md

+++ src/java/gov/health/entity/Person.java
@@ -85,14 +85,16 @@ public class Person implements Serializable {
         this.dateOfDead = dateOfDead;
     }
 
-    public double getAge() {
+    public int getAge() {
         return age;
     }
 
     public void setAge(int age) {
         this.age = age;
     }
+
     
+
     public String getInitials() {
         return initials;
     }
